requirements.txt
pyspark
pandas
pyarrow
nox
pytest
allure-pytest